+++ include/franka/robot_state.h
@@ -92,6 +92,22 @@ struct RobotState {
   std::array<double, 3> F_x_Cee{};  // NOLINT (readability-identifier-naming)
 
   /**
+   * Sum of the mass of the end effector and the external load.
+   */
+  double m_total{};
+
+  /**
+   * Combined rotational inertia matrix of the end effector load and the external load w.r.t.\
+   * center of mass.
+   */
+  std::array<double, 9> I_total{};  // NOLINT (readability-identifier-naming)
+
+  /**
+   * Combined center of mass of the end effector load and the external load w.r.t.\ flange frame.
+   */
+  std::array<double, 3> F_x_Ctotal{};  // NOLINT (readability-identifier-naming)
+
+  /**
    * Elbow configuration.
    *
    * The values of the array are: